refactor: Consolidate intelligence system to use parallel flow only
- Move intelligence-formatter and hitl-review to parallel-intelligence folder
- Remove entire sequential intelligence-gathering system (12 files)
- Update graph to use only parallel intelligence flow
- Simplify architecture by removing duplicate approaches
- All intelligence gathering now uses 4 parallel agents
