--- apps/backend/agents/proposal-generation/nodes/planning/intelligence-gathering/__tests__/integration.test.ts
@@ -1,18 +1,0 @@
-/**
- * Intelligence Gathering Integration Tests
- * Tests for the complete intelligence gathering workflow integration
- * 
- * TODO: These tests need to be rewritten to match the new ReAct pattern architecture:
- * - Test researchAgent with ToolNode pattern
- * - Test intelligenceFormatter instead of synthesis
- * - Test the new routing logic in intelligenceGatheringRouter
- * - Remove references to deleted intelligenceAgent and synthesis nodes
- */
-
-import { describe, it, expect, vi } from "vitest";
-
-describe("Intelligence Gathering Integration", () => {
-  it.skip("needs to be rewritten for new ReAct pattern", () => {
-    expect(true).toBe(true);
-  });
-});
