--- apps/backend/agents/proposal-generation/nodes/planning/intelligence-gathering/progressive-search-framework.ts
@@ -1,1040 +1,0 @@
-// Progressive Search Framework - no external imports needed
-
-/**
- * Progressive Search Framework
- * 
- * A configurable multi-phase search pattern that supports:
- * 1. Discovery - Find relevant sources/pages
- * 2. Extraction - Extract specific entities from discovered sources
- * 3. Deep-dive - Research extracted entities in detail
- */
-
-// ============================================================================
-// Core Types
-// ============================================================================
-
-export interface SearchPhaseResult<T = any> {
-  entities: T[];
-  sources: SearchSource[];
-  metadata: Record<string, any>;
-  quality: QualityAssessment;
-}
-
-export interface SearchSource {
-  url: string;
-  title: string;
-  snippet: string;
-  relevanceScore: number;
-  extractedData?: Record<string, any>;
-}
-
-export interface QualityAssessment {
-  score: number; // 0-1
-  confidence: number; // 0-1
-  completeness: number; // 0-1
-  issues: string[];
-}
-
-export interface ProgressiveSearchConfig<TDiscovery, TExtracted, TEnriched> {
-  // Phase 1: Discovery Configuration
-  discovery: {
-    searchQueries: (context: SearchContext) => string[];
-    sourcePatterns: SourcePattern[];
-    minSources: number;
-    maxSources: number;
-    qualityThreshold: QualityThreshold;
-  };
-
-  // Phase 2: Extraction Configuration
-  extraction: {
-    entityExtractor: EntityExtractor<TDiscovery, TExtracted>;
-    validationRules: ValidationRule<TExtracted>[];
-    deduplicationStrategy: DeduplicationStrategy<TExtracted>;
-    minEntities: number;
-    maxEntities: number;
-  };
-
-  // Phase 3: Deep-dive Configuration
-  deepDive: {
-    searchTemplate: (entity: TExtracted) => DeepDiveQuery[];
-    enrichmentFields: EnrichmentField<TExtracted, TEnriched>[];
-    parallelism: number;
-    timeoutMs: number;
-    qualityThreshold: QualityThreshold;
-  };
-
-  // Global Configuration
-  global: {
-    maxRetries: number;
-    progressCallback?: (phase: SearchPhase, progress: number) => void;
-    errorHandling: ErrorHandlingStrategy;
-  };
-}
-
-// ============================================================================
-// Configuration Building Blocks
-// ============================================================================
-
-export interface SearchContext {
-  domain: string;
-  baseQuery: string;
-  additionalContext: Record<string, any>;
-  previousResults?: SearchPhaseResult[];
-}
-
-export interface SourcePattern {
-  urlPattern?: RegExp;
-  titleKeywords?: string[];
-  contentIndicators?: string[];
-  priority: number;
-}
-
-export interface EntityExtractor<TSource, TEntity> {
-  extract: (source: SearchSource, context: SearchContext) => Promise<TEntity[]>;
-  confidence: (entity: TEntity) => number;
-}
-
-export interface ValidationRule<T> {
-  name: string;
-  validate: (entity: T) => boolean;
-  severity: "error" | "warning";
-}
-
-export interface DeduplicationStrategy<T> {
-  generateKey: (entity: T) => string;
-  mergeDuplicates?: (entities: T[]) => T;
-}
-
-export interface DeepDiveQuery {
-  template: string;
-  requiredFields: string[];
-  optional: boolean;
-}
-
-export interface EnrichmentField<TEntity, TEnriched> {
-  fieldName: keyof TEnriched;
-  extractor: (entity: TEntity, sources: SearchSource[]) => Promise<any>;
-  required: boolean;
-  fallbackValue?: any;
-}
-
-export interface QualityThreshold {
-  minScore: number;
-  minConfidence: number;
-  minCompleteness: number;
-}
-
-export type SearchPhase = "discovery" | "extraction" | "deepDive" | "complete";
-
-export type ErrorHandlingStrategy = "fail" | "continue" | "retry" | "fallback";
-
-// ============================================================================
-// Predefined Configurations
-// ============================================================================
-
-/**
- * Configuration for finding people (current use case)
- */
-export const PEOPLE_SEARCH_CONFIG: ProgressiveSearchConfig<any, Person, PersonProfile> = {
-  discovery: {
-    searchQueries: (ctx) => [
-      `${ctx.domain} team members`,
-      `${ctx.domain} leadership executives`,
-      `${ctx.domain} key personnel staff`,
-      `site:${ctx.domain} about us team`,
-      `site:${ctx.domain} contact directory`
-    ],
-    sourcePatterns: [
-      {
-        urlPattern: /\/(team|about|people|staff|leadership|executives)/i,
-        priority: 10
-      },
-      {
-        titleKeywords: ["team", "staff", "people", "leadership", "about us"],
-        priority: 8
-      }
-    ],
-    minSources: 3,
-    maxSources: 10,
-    qualityThreshold: {
-      minScore: 0.7,
-      minConfidence: 0.6,
-      minCompleteness: 0.5
-    }
-  },
-  extraction: {
-    entityExtractor: {
-      extract: async (source, context) => {
-        // Extract names, titles, departments from source
-        // Implementation would use NLP or pattern matching
-        return [];
-      },
-      confidence: (person) => {
-        // Calculate confidence based on completeness of data
-        return 0.8;
-      }
-    },
-    validationRules: [
-      {
-        name: "hasName",
-        validate: (person) => !!person.name,
-        severity: "error"
-      },
-      {
-        name: "hasRole",
-        validate: (person) => !!person.title,
-        severity: "warning"
-      }
-    ],
-    deduplicationStrategy: {
-      generateKey: (person) => person.name.toLowerCase().replace(/\s+/g, ""),
-      mergeDuplicates: (people) => {
-        // Merge duplicate entries, keeping most complete data
-        return people[0];
-      }
-    },
-    minEntities: 5,
-    maxEntities: 50
-  },
-  deepDive: {
-    searchTemplate: (person) => [
-      {
-        template: `"${person.name}" "${person.company}" linkedin`,
-        requiredFields: ["linkedinUrl", "experience"],
-        optional: false
-      },
-      {
-        template: `"${person.name}" "${person.title}" achievements`,
-        requiredFields: ["achievements"],
-        optional: true
-      }
-    ],
-    enrichmentFields: [
-      {
-        fieldName: "linkedinUrl",
-        extractor: async (person, sources) => {
-          // Extract LinkedIn URL from sources
-          return null;
-        },
-        required: false
-      },
-      {
-        fieldName: "bio",
-        extractor: async (person, sources) => {
-          // Extract and combine bio information
-          return "";
-        },
-        required: false
-      }
-    ],
-    parallelism: 5,
-    timeoutMs: 30000,
-    qualityThreshold: {
-      minScore: 0.6,
-      minConfidence: 0.5,
-      minCompleteness: 0.4
-    }
-  },
-  global: {
-    maxRetries: 3,
-    errorHandling: "continue"
-  }
-};
-
-/**
- * Configuration for finding company products
- */
-export const PRODUCT_SEARCH_CONFIG: ProgressiveSearchConfig<any, Product, ProductDetails> = {
-  discovery: {
-    searchQueries: (ctx) => [
-      `${ctx.domain} products services offerings`,
-      `${ctx.domain} solutions catalog`,
-      `site:${ctx.domain} products`,
-      `site:${ctx.domain} services solutions`
-    ],
-    sourcePatterns: [
-      {
-        urlPattern: /\/(products|services|solutions|offerings)/i,
-        priority: 10
-      },
-      {
-        contentIndicators: ["our products", "we offer", "solutions include"],
-        priority: 7
-      }
-    ],
-    minSources: 2,
-    maxSources: 15,
-    qualityThreshold: {
-      minScore: 0.6,
-      minConfidence: 0.5,
-      minCompleteness: 0.4
-    }
-  },
-  extraction: {
-    entityExtractor: {
-      extract: async (source, context) => {
-        // Extract product names, categories, descriptions
-        return [];
-      },
-      confidence: (product) => {
-        return 0.85;
-      }
-    },
-    validationRules: [
-      {
-        name: "hasName",
-        validate: (product) => !!product.name,
-        severity: "error"
-      },
-      {
-        name: "hasCategory",
-        validate: (product) => !!product.category,
-        severity: "warning"
-      }
-    ],
-    deduplicationStrategy: {
-      generateKey: (product) => `${product.category}:${product.name}`.toLowerCase()
-    },
-    minEntities: 3,
-    maxEntities: 100
-  },
-  deepDive: {
-    searchTemplate: (product) => [
-      {
-        template: `"${product.company}" "${product.name}" features specifications`,
-        requiredFields: ["features", "specifications"],
-        optional: false
-      },
-      {
-        template: `"${product.name}" pricing cost investment`,
-        requiredFields: ["pricing"],
-        optional: true
-      },
-      {
-        template: `"${product.name}" reviews comparison alternatives`,
-        requiredFields: ["reviews", "comparisons"],
-        optional: true
-      }
-    ],
-    enrichmentFields: [
-      {
-        fieldName: "features",
-        extractor: async (product, sources) => {
-          return [];
-        },
-        required: true
-      },
-      {
-        fieldName: "pricing",
-        extractor: async (product, sources) => {
-          return null;
-        },
-        required: false
-      }
-    ],
-    parallelism: 3,
-    timeoutMs: 45000,
-    qualityThreshold: {
-      minScore: 0.7,
-      minConfidence: 0.6,
-      minCompleteness: 0.5
-    }
-  },
-  global: {
-    maxRetries: 2,
-    errorHandling: "continue"
-  }
-};
-
-/**
- * Configuration for finding contract vendors
- */
-export const VENDOR_SEARCH_CONFIG: ProgressiveSearchConfig<any, Vendor, VendorProfile> = {
-  discovery: {
-    searchQueries: (ctx) => [
-      `${ctx.domain} contracts vendors suppliers`,
-      `${ctx.domain} procurement awards`,
-      `site:${ctx.domain} vendor list`,
-      `"${ctx.baseQuery}" contract award`
-    ],
-    sourcePatterns: [
-      {
-        urlPattern: /\/(contracts|vendors|procurement|suppliers)/i,
-        priority: 10
-      },
-      {
-        contentIndicators: ["awarded to", "contract with", "vendor:", "supplier:"],
-        priority: 8
-      }
-    ],
-    minSources: 2,
-    maxSources: 20,
-    qualityThreshold: {
-      minScore: 0.6,
-      minConfidence: 0.5,
-      minCompleteness: 0.4
-    }
-  },
-  extraction: {
-    entityExtractor: {
-      extract: async (source, context) => {
-        // Extract vendor names, contract values, dates
-        return [];
-      },
-      confidence: (vendor) => {
-        return 0.75;
-      }
-    },
-    validationRules: [
-      {
-        name: "hasName",
-        validate: (vendor) => !!vendor.name,
-        severity: "error"
-      },
-      {
-        name: "hasContractInfo",
-        validate: (vendor) => !!vendor.contractValue || !!vendor.contractDate,
-        severity: "warning"
-      }
-    ],
-    deduplicationStrategy: {
-      generateKey: (vendor) => vendor.name.toLowerCase().replace(/[^a-z0-9]/g, "")
-    },
-    minEntities: 5,
-    maxEntities: 200
-  },
-  deepDive: {
-    searchTemplate: (vendor) => [
-      {
-        template: `"${vendor.name}" company profile capabilities`,
-        requiredFields: ["profile", "capabilities"],
-        optional: false
-      },
-      {
-        template: `"${vendor.name}" "${vendor.contractType || 'government'}" contract performance`,
-        requiredFields: ["performance"],
-        optional: true
-      },
-      {
-        template: `"${vendor.name}" certifications qualifications`,
-        requiredFields: ["certifications"],
-        optional: true
-      }
-    ],
-    enrichmentFields: [
-      {
-        fieldName: "capabilities",
-        extractor: async (vendor, sources) => {
-          return [];
-        },
-        required: true
-      },
-      {
-        fieldName: "pastPerformance",
-        extractor: async (vendor, sources) => {
-          return [];
-        },
-        required: false
-      }
-    ],
-    parallelism: 4,
-    timeoutMs: 40000,
-    qualityThreshold: {
-      minScore: 0.65,
-      minConfidence: 0.55,
-      minCompleteness: 0.45
-    }
-  },
-  global: {
-    maxRetries: 3,
-    errorHandling: "continue"
-  }
-};
-
-// ============================================================================
-// Entity Type Definitions
-// ============================================================================
-
-interface Person {
-  name: string;
-  title?: string;
-  department?: string;
-  company: string;
-}
-
-interface PersonProfile extends Person {
-  linkedinUrl?: string;
-  bio?: string;
-  experience?: string[];
-  achievements?: string[];
-  email?: string;
-  phone?: string;
-}
-
-interface Product {
-  name: string;
-  category: string;
-  company: string;
-  description?: string;
-}
-
-interface ProductDetails extends Product {
-  features: string[];
-  specifications?: Record<string, any>;
-  pricing?: {
-    model: string;
-    range?: string;
-  };
-  reviews?: {
-    rating: number;
-    count: number;
-  };
-  comparisons?: Array<{
-    competitor: string;
-    advantages: string[];
-    disadvantages: string[];
-  }>;
-}
-
-interface Vendor {
-  name: string;
-  contractValue?: number;
-  contractDate?: string;
-  contractType?: string;
-}
-
-interface VendorProfile extends Vendor {
-  profile: {
-    description: string;
-    size: string;
-    location: string;
-  };
-  capabilities: string[];
-  certifications?: string[];
-  pastPerformance?: Array<{
-    client: string;
-    project: string;
-    outcome: string;
-  }>;
-}
-
-// ============================================================================
-// Framework Executor
-// ============================================================================
-
-export class ProgressiveSearchExecutor<TDiscovery, TExtracted, TEnriched> {
-  constructor(
-    private config: ProgressiveSearchConfig<TDiscovery, TExtracted, TEnriched>
-  ) {}
-
-  async execute(context: SearchContext): Promise<SearchPhaseResult<TEnriched>> {
-    try {
-      // Phase 1: Discovery
-      this.reportProgress("discovery", 0);
-      const discoveryResults = await this.executeDiscoveryPhase(context);
-      this.reportProgress("discovery", 1);
-
-      // Phase 2: Extraction
-      this.reportProgress("extraction", 0);
-      const extractionResults = await this.executeExtractionPhase(
-        discoveryResults,
-        context
-      );
-      this.reportProgress("extraction", 1);
-
-      // Phase 3: Deep-dive
-      this.reportProgress("deepDive", 0);
-      const enrichedResults = await this.executeDeepDivePhase(
-        extractionResults,
-        context
-      );
-      this.reportProgress("deepDive", 1);
-
-      this.reportProgress("complete", 1);
-      return enrichedResults;
-    } catch (error) {
-      return this.handleError(error);
-    }
-  }
-
-  private async executeDiscoveryPhase(
-    context: SearchContext
-  ): Promise<SearchPhaseResult<TDiscovery>> {
-    const queries = this.config.discovery.searchQueries(context);
-    const sources: SearchSource[] = [];
-
-    // Execute searches in parallel
-    const searchPromises = queries.map(query => this.performSearch(query));
-    const searchResults = await Promise.all(searchPromises);
-
-    // Collect and rank sources
-    for (const results of searchResults) {
-      for (const result of results) {
-        const relevance = this.calculateSourceRelevance(result, context);
-        if (relevance > 0.5) {
-          sources.push({
-            ...result,
-            relevanceScore: relevance
-          });
-        }
-      }
-    }
-
-    // Sort by relevance and limit
-    sources.sort((a, b) => b.relevanceScore - a.relevanceScore);
-    const selectedSources = sources.slice(0, this.config.discovery.maxSources);
-
-    // Assess quality
-    const quality = this.assessDiscoveryQuality(selectedSources);
-
-    return {
-      entities: [] as TDiscovery[],
-      sources: selectedSources,
-      metadata: {
-        queriesExecuted: queries.length,
-        totalSourcesFound: sources.length
-      },
-      quality
-    };
-  }
-
-  private async executeExtractionPhase(
-    discoveryResults: SearchPhaseResult<TDiscovery>,
-    context: SearchContext
-  ): Promise<SearchPhaseResult<TExtracted>> {
-    const extractedEntities: TExtracted[] = [];
-    const { entityExtractor, validationRules, deduplicationStrategy } = this.config.extraction;
-
-    // Extract entities from each source
-    for (const source of discoveryResults.sources) {
-      try {
-        const entities = await entityExtractor.extract(source, context);
-        
-        // Validate entities
-        const validEntities = entities.filter(entity => {
-          const errors = validationRules
-            .filter(rule => rule.severity === "error")
-            .filter(rule => !rule.validate(entity));
-          return errors.length === 0;
-        });
-
-        extractedEntities.push(...validEntities);
-      } catch (error) {
-        console.error(`Extraction failed for source ${source.url}:`, error);
-      }
-    }
-
-    // Deduplicate entities
-    const deduplicatedEntities = this.deduplicateEntities(
-      extractedEntities,
-      deduplicationStrategy
-    );
-
-    // Limit to max entities
-    const selectedEntities = deduplicatedEntities.slice(0, this.config.extraction.maxEntities);
-
-    // Assess quality
-    const quality = this.assessExtractionQuality(selectedEntities, validationRules);
-
-    return {
-      entities: selectedEntities,
-      sources: discoveryResults.sources,
-      metadata: {
-        ...discoveryResults.metadata,
-        totalExtracted: extractedEntities.length,
-        afterDeduplication: deduplicatedEntities.length
-      },
-      quality
-    };
-  }
-
-  private async executeDeepDivePhase(
-    extractionResults: SearchPhaseResult<TExtracted>,
-    context: SearchContext
-  ): Promise<SearchPhaseResult<TEnriched>> {
-    const { searchTemplate, enrichmentFields, parallelism } = this.config.deepDive;
-    const enrichedEntities: TEnriched[] = [];
-
-    // Process entities in batches for parallelism control
-    const batches = this.createBatches(extractionResults.entities, parallelism);
-
-    for (const batch of batches) {
-      const enrichmentPromises = batch.map(async (entity) => {
-        try {
-          // Generate deep-dive queries
-          const queries = searchTemplate(entity);
-          
-          // Execute searches
-          const searchPromises = queries.map(q => 
-            this.performSearch(this.interpolateTemplate(q.template, entity))
-          );
-          const searchResults = await Promise.all(searchPromises);
-          const allSources = searchResults.flat();
-
-          // Enrich entity with extracted data
-          const enrichedEntity = { ...entity } as any;
-          
-          for (const field of enrichmentFields) {
-            try {
-              const value = await field.extractor(entity, allSources);
-              if (value !== undefined || !field.required) {
-                enrichedEntity[field.fieldName] = value ?? field.fallbackValue;
-              }
-            } catch (error) {
-              if (field.required) {
-                throw error;
-              }
-              enrichedEntity[field.fieldName] = field.fallbackValue;
-            }
-          }
-
-          return enrichedEntity as TEnriched;
-        } catch (error) {
-          console.error(`Deep-dive failed for entity:`, entity, error);
-          return null;
-        }
-      });
-
-      const results = await Promise.all(enrichmentPromises);
-      enrichedEntities.push(...results.filter(Boolean) as TEnriched[]);
-      
-      // Report progress
-      const progress = enrichedEntities.length / extractionResults.entities.length;
-      this.reportProgress("deepDive", progress);
-    }
-
-    // Assess quality
-    const quality = this.assessEnrichmentQuality(enrichedEntities, enrichmentFields);
-
-    return {
-      entities: enrichedEntities,
-      sources: extractionResults.sources,
-      metadata: {
-        ...extractionResults.metadata,
-        enrichmentRate: enrichedEntities.length / extractionResults.entities.length
-      },
-      quality
-    };
-  }
-
-  // ============================================================================
-  // Helper Methods
-  // ============================================================================
-
-  private async performSearch(query: string): Promise<SearchSource[]> {
-    // This would integrate with actual search service
-    // For now, return mock data
-    return [];
-  }
-
-  private calculateSourceRelevance(
-    source: SearchSource,
-    context: SearchContext
-  ): number {
-    let score = 0;
-    const patterns = this.config.discovery.sourcePatterns;
-
-    for (const pattern of patterns) {
-      if (pattern.urlPattern && pattern.urlPattern.test(source.url)) {
-        score += pattern.priority * 0.1;
-      }
-      
-      if (pattern.titleKeywords) {
-        const titleLower = source.title.toLowerCase();
-        const matches = pattern.titleKeywords.filter(kw => 
-          titleLower.includes(kw.toLowerCase())
-        );
-        score += (matches.length / pattern.titleKeywords.length) * pattern.priority * 0.05;
-      }
-
-      if (pattern.contentIndicators) {
-        const snippetLower = source.snippet.toLowerCase();
-        const matches = pattern.contentIndicators.filter(indicator =>
-          snippetLower.includes(indicator.toLowerCase())
-        );
-        score += (matches.length / pattern.contentIndicators.length) * pattern.priority * 0.05;
-      }
-    }
-
-    return Math.min(score, 1);
-  }
-
-  private deduplicateEntities(
-    entities: TExtracted[],
-    strategy: DeduplicationStrategy<TExtracted>
-  ): TExtracted[] {
-    const grouped = new Map<string, TExtracted[]>();
-
-    for (const entity of entities) {
-      const key = strategy.generateKey(entity);
-      const existing = grouped.get(key) || [];
-      existing.push(entity);
-      grouped.set(key, existing);
-    }
-
-    const deduplicated: TExtracted[] = [];
-    for (const [key, group] of grouped) {
-      if (group.length === 1) {
-        deduplicated.push(group[0]);
-      } else if (strategy.mergeDuplicates) {
-        deduplicated.push(strategy.mergeDuplicates(group));
-      } else {
-        deduplicated.push(group[0]);
-      }
-    }
-
-    return deduplicated;
-  }
-
-  private createBatches<T>(items: T[], batchSize: number): T[][] {
-    const batches: T[][] = [];
-    for (let i = 0; i < items.length; i += batchSize) {
-      batches.push(items.slice(i, i + batchSize));
-    }
-    return batches;
-  }
-
-  private interpolateTemplate(template: string, entity: any): string {
-    return template.replace(/\$\{(\w+)\}/g, (match, key) => {
-      return entity[key] || match;
-    });
-  }
-
-  private assessDiscoveryQuality(sources: SearchSource[]): QualityAssessment {
-    const avgRelevance = sources.reduce((sum, s) => sum + s.relevanceScore, 0) / sources.length;
-    const hasMinSources = sources.length >= this.config.discovery.minSources;
-
-    return {
-      score: avgRelevance,
-      confidence: hasMinSources ? 0.8 : 0.4,
-      completeness: sources.length / this.config.discovery.maxSources,
-      issues: hasMinSources ? [] : ["Insufficient sources found"]
-    };
-  }
-
-  private assessExtractionQuality(
-    entities: TExtracted[],
-    rules: ValidationRule<TExtracted>[]
-  ): QualityAssessment {
-    const hasMinEntities = entities.length >= this.config.extraction.minEntities;
-    const issues: string[] = [];
-
-    // Check validation warnings
-    for (const entity of entities) {
-      for (const rule of rules) {
-        if (rule.severity === "warning" && !rule.validate(entity)) {
-          issues.push(`Warning: ${rule.name} failed for some entities`);
-          break;
-        }
-      }
-    }
-
-    return {
-      score: hasMinEntities ? 0.8 : 0.4,
-      confidence: 0.7,
-      completeness: entities.length / this.config.extraction.maxEntities,
-      issues: hasMinEntities ? issues : ["Insufficient entities extracted", ...issues]
-    };
-  }
-
-  private assessEnrichmentQuality(
-    entities: TEnriched[],
-    fields: EnrichmentField<TExtracted, TEnriched>[]
-  ): QualityAssessment {
-    const requiredFields = fields.filter(f => f.required);
-    let completenessSum = 0;
-
-    for (const entity of entities) {
-      const filledRequired = requiredFields.filter(f => 
-        entity[f.fieldName] !== undefined && entity[f.fieldName] !== null
-      );
-      completenessSum += filledRequired.length / requiredFields.length;
-    }
-
-    const avgCompleteness = entities.length > 0 ? completenessSum / entities.length : 0;
-
-    return {
-      score: avgCompleteness,
-      confidence: 0.75,
-      completeness: avgCompleteness,
-      issues: avgCompleteness < 0.8 ? ["Some required fields missing"] : []
-    };
-  }
-
-  private reportProgress(phase: SearchPhase, progress: number): void {
-    if (this.config.global.progressCallback) {
-      this.config.global.progressCallback(phase, progress);
-    }
-  }
-
-  private handleError(error: any): SearchPhaseResult<TEnriched> {
-    const strategy = this.config.global.errorHandling;
-    
-    if (strategy === "fail") {
-      throw error;
-    }
-
-    return {
-      entities: [],
-      sources: [],
-      metadata: { error: error.message },
-      quality: {
-        score: 0,
-        confidence: 0,
-        completeness: 0,
-        issues: [`Error: ${error.message}`]
-      }
-    };
-  }
-}
-
-// ============================================================================
-// Factory Functions
-// ============================================================================
-
-export function createPeopleSearcher(
-  overrides?: Partial<ProgressiveSearchConfig<any, Person, PersonProfile>>
-): ProgressiveSearchExecutor<any, Person, PersonProfile> {
-  const config = { ...PEOPLE_SEARCH_CONFIG, ...overrides };
-  return new ProgressiveSearchExecutor(config);
-}
-
-export function createProductSearcher(
-  overrides?: Partial<ProgressiveSearchConfig<any, Product, ProductDetails>>
-): ProgressiveSearchExecutor<any, Product, ProductDetails> {
-  const config = { ...PRODUCT_SEARCH_CONFIG, ...overrides };
-  return new ProgressiveSearchExecutor(config);
-}
-
-export function createVendorSearcher(
-  overrides?: Partial<ProgressiveSearchConfig<any, Vendor, VendorProfile>>
-): ProgressiveSearchExecutor<any, Vendor, VendorProfile> {
-  const config = { ...VENDOR_SEARCH_CONFIG, ...overrides };
-  return new ProgressiveSearchExecutor(config);
-}
-
-// ============================================================================
-// Usage Examples
-// ============================================================================
-
-/*
-// Example 1: Search for people at a company
-const peopleSearcher = createPeopleSearcher({
-  discovery: {
-    maxSources: 20 // Increase source limit
-  },
-  deepDive: {
-    parallelism: 10 // Increase parallelism for faster enrichment
-  }
-});
-
-const peopleResults = await peopleSearcher.execute({
-  domain: "example.com",
-  baseQuery: "Example Corp",
-  additionalContext: {
-    focusRoles: ["engineering", "product"]
-  }
-});
-
-// Example 2: Search for products with custom extraction
-const productSearcher = createProductSearcher({
-  extraction: {
-    entityExtractor: {
-      extract: async (source, context) => {
-        // Custom extraction logic
-        return customProductExtractor(source);
-      },
-      confidence: (product) => calculateProductConfidence(product)
-    }
-  }
-});
-
-const productResults = await productSearcher.execute({
-  domain: "techcompany.com",
-  baseQuery: "TechCompany",
-  additionalContext: {
-    productCategories: ["software", "services"]
-  }
-});
-
-// Example 3: Create a custom search pattern
-const customConfig: ProgressiveSearchConfig<any, NewsItem, EnrichedNewsItem> = {
-  discovery: {
-    searchQueries: (ctx) => [
-      `"${ctx.baseQuery}" news announcements`,
-      `"${ctx.baseQuery}" press releases`
-    ],
-    sourcePatterns: [
-      {
-        urlPattern: /\/(news|press|announcements)/i,
-        priority: 10
-      }
-    ],
-    minSources: 5,
-    maxSources: 30,
-    qualityThreshold: {
-      minScore: 0.6,
-      minConfidence: 0.5,
-      minCompleteness: 0.4
-    }
-  },
-  extraction: {
-    entityExtractor: {
-      extract: async (source, context) => {
-        // Extract news items
-        return extractNewsItems(source);
-      },
-      confidence: (item) => 0.9
-    },
-    validationRules: [
-      {
-        name: "hasDate",
-        validate: (item) => !!item.date,
-        severity: "error"
-      }
-    ],
-    deduplicationStrategy: {
-      generateKey: (item) => item.title.toLowerCase()
-    },
-    minEntities: 10,
-    maxEntities: 100
-  },
-  deepDive: {
-    searchTemplate: (item) => [
-      {
-        template: `"${item.title}" impact analysis`,
-        requiredFields: ["impact"],
-        optional: true
-      }
-    ],
-    enrichmentFields: [
-      {
-        fieldName: "sentiment",
-        extractor: async (item, sources) => analyzeSentiment(sources),
-        required: false
-      }
-    ],
-    parallelism: 5,
-    timeoutMs: 30000,
-    qualityThreshold: {
-      minScore: 0.7,
-      minConfidence: 0.6,
-      minCompleteness: 0.5
-    }
-  },
-  global: {
-    maxRetries: 2,
-    errorHandling: "continue",
-    progressCallback: (phase, progress) => {
-      console.log(`${phase}: ${Math.round(progress * 100)}%`);
-    }
-  }
-};
-
-const newsSearcher = new ProgressiveSearchExecutor(customConfig);
-const newsResults = await newsSearcher.execute({
-  domain: "company.com",
-  baseQuery: "Company Name",
-  additionalContext: {}
-});
-*/
