+++ apps/backend/agents/proposal-generation/nodes/planning/parallel-intelligence/intelligence-formatter.ts
